Removed DragWidget combining it into PieWidget
