+++ shortcut_composer/templates/pie_menu.py
@@ -102,7 +102,7 @@ class PieMenu(ComplexAction, Generic[T]):
         self._labels = self._create_labels(values)
         self._style.adapt_to_item_amount(len(self._labels))
 
-        self._pie_manager = PieManager(PieWidget(self._labels, self._style))
+        self._pie_manager = PieManager(PieWidget(self._style, self._labels))
 
     def on_key_press(self) -> None:
         """Show widget under mouse and start manager which repaints it."""
